Update using-sqip example to gatsby v2
