+++ examples/using-sqip/src/pages/index.js
@@ -1,7 +1,11 @@
 import React from 'react'
+import PropTypes from 'prop-types'
+import { graphql } from 'gatsby'
 
-const IndexPage = () => (
-  <div>
+import Layout from '../components/layout.js'
+
+const IndexPage = ({ data }) => (
+  <Layout data={data}>
     <h1>Gatsby SQIP Example</h1>
     <blockquote>
       <p>
@@ -14,17 +18,20 @@ const IndexPage = () => (
       <br /> An algorithm calculates a primitive representation of your images
       based on simple shapes like circles, ellipses, triangles and more. These
       will be embedded in your initial HTML payload. This will help your users
-      to get a feeling of how the pictures will look like, even{` `}
+      to get a feeling of how the pictures will look like, even
+      {` `}
       <strong>before</strong> they got loaded by their (probably) slow
       connection.
     </p>
     <h2>Configuration</h2>
     <p>
       For an up to date list of possible configuration options, please check out
-      the{` `}
+      the
+      {` `}
       <a href="https://www.gatsbyjs.org/packages/gatsby-transformer-sqip/">
         plugins readme
-      </a>.
+      </a>
+      .
     </p>
     <h3>Configuration and file size recommendations:</h3>
     <p>
@@ -42,7 +49,8 @@ const IndexPage = () => (
     </ul>
     <p>
       Generally: <strong>Keep it as small as possible</strong> and test the
-      impact of your image previews via{` `}
+      impact of your image previews via
+      {` `}
       <a href="https://www.webpagetest.org/">webpagetest.org</a> on a 3G
       connection.
     </p>
@@ -50,11 +58,11 @@ const IndexPage = () => (
     <p>Getting the data via GraphQL:</p>
     <pre>
       <code>{`image {
-  sqip(numberOfPrimitives: 12, blur: 12, width: 256, height: 256),
-  sizes(maxWidth: 400, maxHeight: 400) {
-   ...GatsbyImageSharpSizes_noBase64
-  }
-}`}</code>
+    sqip(numberOfPrimitives: 12, blur: 12, width: 256, height: 256),
+    sizes(maxWidth: 400, maxHeight: 400) {
+    ...GatsbyImageSharpSizes_noBase64
+    }
+  }`}</code>
     </pre>
     <p>
       <strong>Hint:</strong> Make sure to set the same aspect ratio for sqip and
@@ -142,10 +150,64 @@ const IndexPage = () => (
     <h2>Copyright notice</h2>
     <p>
       The used images belong to their photographer, get more details by clicking
-      on them. The background image was taken by{` `}
+      on them. The background image was taken by
+      {` `}
       <a href="https://unsplash.com/photos/ljHAKAnYs4I">Neven Krcmarek</a>.
     </p>
-  </div>
+  </Layout>
 )
 
 export default IndexPage
+
+IndexPage.propTypes = {
+  data: PropTypes.object,
+}
+
+export const query = graphql`
+  query {
+    images: allFile(
+      filter: { sourceInstanceName: { eq: "images" }, ext: { eq: ".jpg" } }
+    ) {
+      edges {
+        node {
+          publicURL
+          name
+          childImageSharp {
+            fluid(maxWidth: 400, maxHeight: 400) {
+              ...GatsbyImageSharpFluid_noBase64
+            }
+            sqip(
+              # Make sure to keep the same aspect ratio when cropping
+              # With 256px as maximum dimension is the perfect value to speed up the process
+              width: 256
+              height: 256
+              numberOfPrimitives: 15
+              blur: 8
+              mode: 1
+            ) {
+              dataURI
+            }
+          }
+        }
+      }
+    }
+    background: allFile(
+      filter: { sourceInstanceName: { eq: "background" }, ext: { eq: ".jpg" } }
+    ) {
+      edges {
+        node {
+          publicURL
+          name
+          childImageSharp {
+            fluid(maxWidth: 4000) {
+              ...GatsbyImageSharpFluid_noBase64
+            }
+            sqip(numberOfPrimitives: 160, blur: 0) {
+              dataURI
+            }
+          }
+        }
+      }
+    }
+  }
+`
